4000 Rename KnowledgeBaseTagServiceImpl -> KnowledgeBaseTagFacadeImpl, as TagService is used inside
